cleaning FormationController

+++ src/main/java/com/resume/web/controller/FormationController.java
@@ -1,19 +1,12 @@
 package com.resume.web.controller;
 
-import com.resume.converter.ConverterHelper;
 import com.resume.dco.FormationDco;
 import com.resume.dto.FormationDto;
-import com.resume.model.Formation;
-import com.resume.model.Project;
-import com.resume.model.enums.FormationTypeEnum;
-import com.resume.repository.FormationRepository;
-import com.resume.repository.LocationRepository;
 import com.resume.web.exceptions.NoContentException;
 import com.resume.web.exceptions.UnexpectedCompetenceException;
 import com.resume.web.exceptions.UnexpectedFormationException;
 import com.resume.web.serviceImpl.FormationServiceImpl;
 import io.swagger.annotations.ApiOperation;
-import org.modelmapper.ModelMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
@@ -27,19 +20,10 @@ public class FormationController {
 
     public static final Logger logger = LoggerFactory.getLogger(FormationController.class);
 
-    private FormationServiceImpl formationService;
+    private final FormationServiceImpl formationService;
 
-    private LocationRepository locationRepository;
-
-    private FormationRepository formationRepository;
-
-    private ModelMapper modelMapper;
-
-    public FormationController(FormationServiceImpl formationService, LocationRepository locationRepository, FormationRepository formationRepository, ModelMapper modelMapper) {
+    public FormationController(FormationServiceImpl formationService) {
         this.formationService = formationService;
-        this.locationRepository = locationRepository;
-        this.formationRepository = formationRepository;
-        this.modelMapper = modelMapper;
     }
 
     @CrossOrigin()
@@ -53,7 +37,7 @@ public class FormationController {
     @ApiOperation(value = "Post a formation")
     @PostMapping(value = "/Formations")
     public ResponseEntity<FormationDto> post(@Valid @RequestBody FormationDco formationDco) {
-        return ResponseEntity.ok().body(formationService.postFormation(ConverterHelper.convertToEntity(formationDco, modelMapper, locationRepository)));
+        return ResponseEntity.ok().body(formationService.postFormation(formationDco));
     }
 
     @CrossOrigin()
@@ -61,20 +45,14 @@ public class FormationController {
     @PutMapping(value = "/Formations")
     public ResponseEntity<FormationDto> put(@Valid @RequestBody FormationDco formationDco) {
         logger.info("Updating formation with id {}", formationDco.getFormationId());
-        Formation currentFormation = formationRepository.findByFormationId(formationDco.getFormationId());
+        FormationDto updatedFormation = formationService.putFormation(formationDco);
 
-        if (currentFormation == null) {
+        if (updatedFormation == null) {
             logger.error("Unable to update. Formation with id {} not found.", formationDco.getFormationId());
-            throw new UnexpectedFormationException("Unable to upate. Organisation with id " + formationDco.getFormationId() + " not found.");
+            throw new UnexpectedFormationException("Unable to update. Organisation with id " + formationDco.getFormationId() + " not found.");
         }
 
-        currentFormation.setStartDate(formationDco.getStartDate());
-        currentFormation.setEndDate(formationDco.getEndDate());
-        currentFormation.setFormationTitle(formationDco.getFormationTitle());
-        currentFormation.setFormationType(FormationTypeEnum.valueOf(formationDco.getFormationType()));
-        currentFormation.setLocation(locationRepository.findByLocationId(formationDco.getLocationId()));
-
-        return ResponseEntity.ok().body(formationService.putFormation(currentFormation));
+        return ResponseEntity.ok().body(updatedFormation);
     }
 
     @CrossOrigin()
@@ -92,9 +70,14 @@ public class FormationController {
     @CrossOrigin()
     @ApiOperation(value = "Delete a formation")
     @DeleteMapping(value = "/Formations/{formationId}")
-    public ResponseEntity<Formation> delete(@PathVariable Integer formationId) {
+    public ResponseEntity<FormationDto> delete(@PathVariable Integer formationId) {
         logger.info("Deleting formation with id {}", formationId);
-        Formation formationToDelete = formationRepository.findByFormationId(formationId);
+        FormationDto formationToDelete = formationService.getFormation(formationId);
+
+        if(formationToDelete == null) {
+            throw new UnexpectedCompetenceException("Unable to delete. Formation with id " + formationId + " not found.");
+        }
+
         Integer deletedFormation  = formationService.deleteFormation(formationId);
 
         if(deletedFormation == 0) {
